refactor: moving builtarget into builders/golang
Its only used in the golang builder anyway.

first step towards #5450
