+++ internal/builders/golang/targets_test.go
@@ -1,4 +1,4 @@
-package buildtarget
+package golang
 
 import (
 	"fmt"
@@ -111,7 +111,7 @@ func TestAllBuildTargets(t *testing.T) {
 	}
 
 	t.Run("go 1.18", func(t *testing.T) {
-		result, err := List(build)
+		result, err := listTargets(build)
 		require.NoError(t, err)
 		require.Equal(t, []string{
 			"linux_386_softfloat",
@@ -163,7 +163,7 @@ func TestAllBuildTargets(t *testing.T) {
 	})
 
 	t.Run("invalid goos", func(t *testing.T) {
-		_, err := List(config.Build{
+		_, err := listTargets(config.Build{
 			Goos:    []string{"invalid"},
 			Goarch:  []string{"amd64"},
 			Goamd64: []string{"v2"},
@@ -172,7 +172,7 @@ func TestAllBuildTargets(t *testing.T) {
 	})
 
 	t.Run("invalid goarch", func(t *testing.T) {
-		_, err := List(config.Build{
+		_, err := listTargets(config.Build{
 			Goos:   []string{"linux"},
 			Goarch: []string{"invalid"},
 		})
@@ -180,7 +180,7 @@ func TestAllBuildTargets(t *testing.T) {
 	})
 
 	t.Run("invalid goarm", func(t *testing.T) {
-		_, err := List(config.Build{
+		_, err := listTargets(config.Build{
 			Goos:   []string{"linux"},
 			Goarch: []string{"arm"},
 			Goarm:  []string{"invalid"},
@@ -189,7 +189,7 @@ func TestAllBuildTargets(t *testing.T) {
 	})
 
 	t.Run("invalid gomips", func(t *testing.T) {
-		_, err := List(config.Build{
+		_, err := listTargets(config.Build{
 			Goos:   []string{"linux"},
 			Goarch: []string{"mips"},
 			Gomips: []string{"invalid"},
@@ -198,7 +198,7 @@ func TestAllBuildTargets(t *testing.T) {
 	})
 
 	t.Run("invalid goamd64", func(t *testing.T) {
-		_, err := List(config.Build{
+		_, err := listTargets(config.Build{
 			Goos:    []string{"linux"},
 			Goarch:  []string{"amd64"},
 			Goamd64: []string{"invalid"},
@@ -267,14 +267,14 @@ func TestGoosGoarchCombos(t *testing.T) {
 	}
 	for _, p := range platforms {
 		t.Run(fmt.Sprintf("%v %v valid=%v", p.os, p.arch, p.valid), func(t *testing.T) {
-			require.Equal(t, p.valid, valid(target{os: p.os, arch: p.arch}))
+			require.Equal(t, p.valid, valid(Target{Goos: p.os, Goarch: p.arch}))
 		})
 	}
 }
 
-func TestList(t *testing.T) {
+func TestListTargets(t *testing.T) {
 	t.Run("success", func(t *testing.T) {
-		targets, err := List(config.Build{
+		targets, err := listTargets(config.Build{
 			Goos:    []string{"linux"},
 			Goarch:  []string{"amd64"},
 			Goamd64: []string{"v2"},
@@ -285,7 +285,7 @@ func TestList(t *testing.T) {
 	})
 
 	t.Run("success with dir", func(t *testing.T) {
-		targets, err := List(config.Build{
+		targets, err := listTargets(config.Build{
 			Goos:    []string{"linux"},
 			Goarch:  []string{"amd64"},
 			Goamd64: []string{"v2"},
